s3-agent-local/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.34.0
bedrock-agentcore>=0.1.21
moto[s3]>=5.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
